Fixed evaluation bug
The initial minimax code does not evaluate properly because of two things:

1. the depth checking was placed before the player winner checks in the
minimax function, causing the function to return zero scores at leaf
nodes even if there is a winning player.

2. the initial std::map key-pair value stores the ttt::Move type and
there seems to be a problem in storing or retreiving it in the std::map,
causing wrong evaluation scores (this might be because I did not declare
a proper move and copy constructor [but I doubt it]). anyways the problem
was fixed when I optimized the size of the std::map by changing the
key-pair's "pair" type from "ttt::Move" to float.